+++ src/pages/asset/main/AccountOverviewCard.vue
@@ -1,5 +1,6 @@
 <template>
   <div class="overview-card">
+    <!-- 타이틀 -->
     <div class="title-row" @click="goToAccountTab">
       <h3>계좌 현황</h3>
       <div class="arrow">
@@ -11,64 +12,23 @@
       </div>
     </div>
 
-    <!-- 계좌 목록: 최대 3개 표시 -->
-    <div
-      v-for="(account, index) in accounts.slice(0, 3)"
-      :key="index"
-      class="account-item"
-    >
-      <img
-        :src="getBankLogoByCode(account.bankCode)"
-        :alt="`${getBankName(account.bankCode)} 로고`"
-        class="bank-logo"
-      />
-      <div class="info">
-        <p class="name">
-          {{ getBankName(account.bankCode) }} {{ account.accountName }}
-        </p>
-        <p class="number">{{ formatAccountNumber(account.accountNumber) }}</p>
-      </div>
-      <div class="balance">
-        <p class="amount">{{ formatWon(account.balance) }}</p>
-        <p class="type">{{ formatType(account.accountType) }}</p>
-      </div>
-    </div>
+    <!-- 계좌 리스트 (최대 3개 표시) -->
+    <AccountList :accounts="accounts.slice(0, 3)" />
   </div>
 </template>
 
 <script setup>
 import { ref, onMounted } from 'vue';
-import { useRouter } from 'vue-router';
-
-//계좌
+import AccountList from '../account/AccountList.vue';
 import rawAccounts from '@/assets/data/accounts.json';
-import { getBankLogoByCode } from '@/assets/utils/bankLogoMap.js';
-import { getBankName } from '@/assets/utils/bankCodeMap.js';
 
-const router = useRouter();
 const emit = defineEmits(['switchTab']);
 const accounts = ref([]);
 
-function goToAccountTab() {
+// 내부 탭 전환: 계좌탭 이동
+const goToAccountTab = () => {
   emit('switchTab', '계좌');
-}
-
-function formatWon(value) {
-  return `${value.toLocaleString()}원`;
-}
-
-function formatAccountNumber(number) {
-  return number.replace(/(\d{3})(\d{2,3})(\d{4})/, '$1-$2-$3');
-}
-
-function formatType(type) {
-  const map = {
-    11: '입출금통장',
-    12: '적금',
-    13: '예금',
-  };
-  return map[type] || '기타';
-}
+};
 
 onMounted(() => {
   accounts.value = rawAccounts;
@@ -97,53 +57,8 @@ onMounted(() => {
   color: var(--text-title, #111827);
 }
 
-.account-item {
-  display: flex;
-  align-items: center;
-  background-color: var(--input-bg-2, #f8fafc);
-  padding: 0.75rem;
-  border-radius: 0.75rem;
-  margin-bottom: 0.75rem;
-}
-
 .arrow-icon {
   width: 1rem;
   height: 1rem;
-}
-
-.bank-logo {
-  width: 2.5rem;
-  height: 2.5rem;
-  object-fit: contain;
-  margin-right: 0.75rem;
-}
-
-.info {
-  flex: 1;
-}
-
-.name {
-  font-weight: 600;
-  font-size: 0.875rem;
-}
-
-.number {
-  font-size: 0.75rem;
-  color: var(--text-lightgray);
-}
-
-.balance {
-  text-align: right;
-}
-
-.amount {
-  font-weight: 600;
-  font-size: 1rem;
-  color: var(--base-blue-dark);
-}
-
-.type {
-  font-size: 0.8125rem;
-  color: var(--text-lightgray);
 }
 </style>
